refine master client
